pyproject.toml
[tool.poetry]
name = "ct-backend"
version = "0.1.0"
description = "Creator Tools Backend — FastAPI + Poetry"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111"
uvicorn = { extras = ["standard"], version = "^0.30" }
sqlalchemy = "^2.0"
pydantic-settings = "^2.0"
alembic = "^1.13"
python-dotenv = "^1.0"
# Auth
python-jose = { extras = ["cryptography"], version = "^3.3" }
passlib = { extras = ["bcrypt"], version = "^1.7" }
# Logging
loguru = "^0.7"
# Rate limiting
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8"
pytest = "^8.0"
pytest-asyncio = "^0.23"
httpx = "^0.27"
black = "^24.0"
isort = "^5.13"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
